Update HOTS
Instead of jumping directly to the requested build or wiki page, the program created an instance of the Hero class to check if the requested hero existed first, which resulted in an extended wait time for the page to appear. Rather than making an instance, the script now checks if the requested page alone exists, then passes it along. 

I also added one word keywords and automatic name edits for when inputting a hero's name, eg. if you input "butcher" or "hammer", it will auto change it to "the-butcher" or "sgt-hammer" for the builds website.